Fix Intel build (I hope).
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@4379 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ pde/src/solver/common/NonlinearElasticityAssembler.hpp
@@ -56,7 +56,7 @@ friend class TestNonlinearElasticityAssembler;
 protected:
     static const unsigned NUM_VERTICES_PER_ELEMENT = DIM+1;
     static const unsigned NUM_NODES_PER_ELEMENT = (DIM+1)*(DIM+2)/2; // assuming quadratic
-    static const unsigned STENCIL_SIZE = DIM*NUM_NODES_PER_ELEMENT + NUM_VERTICES_PER_ELEMENT; 
+    static const size_t STENCIL_SIZE = DIM*NUM_NODES_PER_ELEMENT + NUM_VERTICES_PER_ELEMENT; 
     static const unsigned NUM_NODES_PER_BOUNDARY_ELEMENT = DIM*(DIM+1)/2;
     static const unsigned BOUNDARY_STENCIL_SIZE = DIM*NUM_NODES_PER_BOUNDARY_ELEMENT + DIM;
